Add getters for protected members

+++ include/sonify/MapTemplate.hpp
@@ -16,20 +16,20 @@ public:
     virtual const char *name() const                               = 0;
     virtual std::vector<short> mapping(const std::vector<Pixel> &) = 0;
 
-    inline int minFreq() const noexcept { return _min_freq; }
-    inline int maxFreq() const noexcept { return _max_freq; }
-    inline int sampleRate() const noexcept { return _sample_rate; }
+    inline float minFreq() const noexcept { return _min_freq; }
+    inline float maxFreq() const noexcept { return _max_freq; }
+    inline float sampleRate() const noexcept { return _sample_rate; }
     inline FreqMapFunc freqMapper() const noexcept { return freq_map; }
 
-    inline void setMinFreq(int f) noexcept { _min_freq = f; }
-    inline void setMaxFreq(int f) noexcept { _max_freq = f; }
-    inline void setSampleRate(int f) noexcept { _sample_rate = f; }
+    inline void setMinFreq(float f) noexcept { _min_freq = f; }
+    inline void setMaxFreq(float f) noexcept { _max_freq = f; }
+    inline void setSampleRate(float f) noexcept { _sample_rate = f; }
     inline void setFreqMap(FreqMapFunc f) noexcept { freq_map = f; }
 
 protected:
 
     FreqMapFunc freq_map{ utils::LinearMap };
-    int _min_freq{ 0 }, _max_freq{ 20000 }, _sample_rate{ 44100 };
+    float _min_freq{ 0.0f }, _max_freq{ 20000.0f }, _sample_rate{ 44100.0f };
 };
 
 // Factory function for plugins
